Create UserModule. I should technically move some GraphQL stuff over from AccountModule, but meh...
