+++ FarmMaster.sln
@@ -35,7 +35,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FarmMaster.Module.TagHelper
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataAccess", "DataAccess", "{2D1E68CF-FBF4-48BB-A85A-EF2BE9FAD8A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQLModule", "GraphQLModule\GraphQLModule.csproj", "{2CA5CCDC-04E0-4F42-9477-28302E8B34DD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQLModule", "GraphQLModule\GraphQLModule.csproj", "{2CA5CCDC-04E0-4F42-9477-28302E8B34DD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserModule", "UserModule\UserModule.csproj", "{D989C608-D8F9-401A-AAC0-C3969BD65CAF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -91,6 +93,10 @@ Global
 		{2CA5CCDC-04E0-4F42-9477-28302E8B34DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2CA5CCDC-04E0-4F42-9477-28302E8B34DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2CA5CCDC-04E0-4F42-9477-28302E8B34DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D989C608-D8F9-401A-AAC0-C3969BD65CAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D989C608-D8F9-401A-AAC0-C3969BD65CAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D989C608-D8F9-401A-AAC0-C3969BD65CAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D989C608-D8F9-401A-AAC0-C3969BD65CAF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -108,6 +114,7 @@ Global
 		{A86242E0-53C8-4356-A5E6-86CEA830D895} = {8A2E8B38-5CC7-4EDD-958F-874E6B092B37}
 		{C9BD2E21-6E6D-4746-8BBE-4AC24C566C45} = {8A2E8B38-5CC7-4EDD-958F-874E6B092B37}
 		{2CA5CCDC-04E0-4F42-9477-28302E8B34DD} = {7E337641-E563-4798-B9E0-9A4794AB847E}
+		{D989C608-D8F9-401A-AAC0-C3969BD65CAF} = {7E337641-E563-4798-B9E0-9A4794AB847E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A8B24A22-12A2-4331-BE7D-763A026063E2}
