+++ UserModule/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,0 +1,7 @@
+﻿@using DataAccess
+@using UserModule.Models
+@using Microsoft.AspNetCore.Identity
+@using FarmMaster.Module.Core.Models
+
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, FarmMaster.Module.TagHelpers
